+++ src/record/rm_defs.h
@@ -100,8 +100,8 @@ struct RmRecord {
     T import(int offset, int len = 0) const {
         if constexpr (std::is_same_v<T, int>) {
             return *reinterpret_cast<int*>(data + offset);
-        } else if constexpr (std::is_same_v<T, double>) {
-            return *reinterpret_cast<double*>(data + offset);
+        } else if constexpr (std::is_same_v<T, float>) {
+            return *reinterpret_cast<float*>(data + offset);
         } else if constexpr (std::is_same_v<T, std::string>) {
             return std::string(data + offset, len);
         }
